+++ frontend/src/app/components/containers/complaints/complaint-map.tsx
@@ -6,11 +6,7 @@ import { ComplaintFilterContext } from "../../../providers/complaint-filter-prov
 import { ComplaintFilters } from "../../../types/complaints/complaint-filters/complaint-filters";
 import { ComplaintRequestPayload } from "../../../types/complaints/complaint-filters/complaint-reauest-payload";
 import LeafletMapWithMultiplePoints from "../../mapping/leaflet-map-with-multiple-points";
-import {
-  getComplaintsOnMap,
-  setComplaintsOnMap,
-  selectComplaintLocations,
-} from "../../../store/reducers/complaints";
+import { getComplaintsOnMap, selectComplaintLocations, setComplaintsOnMap } from "../../../store/reducers/complaintLocations";
 
 type Props = {
   type: string;
@@ -85,5 +81,9 @@ export const ComplaintMap: FC<Props> = ({ type }) => {
     };
   }, []);
 
-  return <LeafletMapWithMultiplePoints markers={coordinatesArray} />;
+  useEffect(() => {
+    //alert('test10');
+  }, [coordinatesArray]);
+
+  return <LeafletMapWithMultiplePoints complaint_type={type} markers={coordinatesArray} />;
 };
