+++ frontend/src/assets/sass/complaint.scss
@@ -211,11 +211,15 @@
 }
 
 .comp-status-badge-open {
-  background-color: $bc-gov-link-blue;
+  background-color: $green-100-80;
+  color: $green-100;
+  line-height: normal;
 }
 
 .comp-status-badge-closed {
-  background-color: $bc-gov-danger;
+  background-color: $red-300-80;
+  color: $red-300;
+  line-height: normal;
 }
 
 .comp-complaint-details-block {
