+++ frontend/src/assets/sass/maps.scss
@@ -12,32 +12,37 @@
     z-index: 1;
 }
 
-.map-comp-popup {
-    height: hug(400px);
-    width: 370px;
+
+.leaflet-popup-content-wrapper {
+    display: flex;
+    justify-content: center;
+}
+
+.leaflet-popup-content {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: auto !important;
+    /* Set width to auto */
+    max-width: 100%;
+    /* Ensure the content does not exceed the available width */
 }
 
 .map-comp-summary-popup-container {
-
-    width: 330px;
     display: flex;
+    gap: 24px;
     flex-direction: column;
-    justify-content: flex-start; /* Align items to the left */
-
+    justify-content: flex-start;
+    /* Align items to the left */
 }
 
-.map-comp-summary-popup-details {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    font-size: 16px;
+.map-comp-summary-popup-subheading-item {
+    white-space: nowrap;
 }
 
 .map-comp-summary-popup-header {
     display: flex;
     justify-content: space-between;
-    margin-bottom: 5px;
 }
 
 .map-comp-summary-popup-header .complaint-identifier {
@@ -54,9 +59,32 @@
 
 .map-comp-summary-popup-subheading {
     display: flex;
-    justify-content: flex-start; /* Align items to the left */
+    justify-content: space-between;
     gap: 12px;
     color: $white;
+}
+
+.map-comp-summary-popup-subheading:last-child {
+    margin-left: auto;
+}
+
+.map-comp-summary-popup-details {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    font-size: 16px;
+
+    .comp-details-content {
+        display: flex;
+        flex-direction: column;
+        gap: 12px;
+    }
+
+    .comp-details-content label {
+        margin-right: 10px;
+        color: $gray-600;
+        ;
+    }
 }
 
 .map-comp-nature-of-complaint {
@@ -73,8 +101,8 @@
 /* Each row within the other-div */
 .map-comp-summary-popup-details-section .row {
     display: flex;
+    flex-direction: row;
     justify-content: space-between;
-    margin-bottom: 10px;
 }
 
 .map-comp-summary-popup-subheading .comp-box-conflict-type {
@@ -86,18 +114,16 @@
     font-weight: bold;
     line-height: 16px;
     display: flex;
-  }
+}
 
-  .map-comp-summary-popup-subheading .comp-box-conflict-type.hwcr-conflict-type {
+.map-comp-summary-popup-subheading .comp-box-conflict-type.hwcr-conflict-type {
     background: $bc-standard-03;
     margin: 0px !important
-  }
-  .map-comp-summary-popup-subheading .comp-box-conflict-type.allegation-conflict-type {
+}
+
+.map-comp-summary-popup-subheading .comp-box-conflict-type.allegation-conflict-type {
     background: $bc-standard-05;
-  }
-
-
-.map-comp-summary-popup-button {}
+}
 
 .popup-visible {
     visibility: visible;
